More sane reset_props code

+++ gamemodes/jazztronauts/gamemode/lib/sh_unlocks.lua
@@ -30,7 +30,8 @@ function ClearAll()
 end
 
 function ClearPlayer( list_name, steamid64 )
-	if not list_name or not steamid64 then return false end
+
+	if not list_name then return false end
 
 	if CLIENT then
 		local list = unlock_lists[list_name]
@@ -38,17 +39,20 @@ function ClearPlayer( list_name, steamid64 )
 		return true
 	end
 
+	if not steamid64 then return false end
+
 	local result = sql.Query( ("DELETE FROM %s WHERE steamid = '%s'"):format(
 		unlock_lists[list_name],
 		steamid64 ) )
 
-	if result == false then
+	if false == result then
 		print("ERROR: " .. tostring( sql.LastError() ) )
 		return false
 	end
 
 	print("Successfully wiped " .. list_name .. " for " .. steamid64)
 	return true
+
 end
 
 function Register( list_name )
@@ -319,7 +323,7 @@ end )
 local ResetHelpMsg = "Resets your collected props. This does not affect the sandbox spawnmenu, only snatched props."
 	.. "\nDoing this can possibly reduce lag when opening the spawnmenu or claiming props in the hub."
 	.. '\nTo clear yourself, type "jazz_reset_props self". As an admin, to clear it for everyone, type "all" instead.'
-	.. "\nAdmins can also enter a player's SteamID64. They don't need to be online."
+	.. "\nAdmins can also enter a player's SteamID64. The player being reset doesn't need to be online."
 concommand.Add( "jazz_reset_props", function( ply, _, args )
 	local arg = args[1]
 
@@ -328,19 +332,23 @@ concommand.Add( "jazz_reset_props", function( ply, _, args )
 		return
 	end
 
-	if arg == "all" and ply:IsAdmin() then
+	if ply:IsAdmin() and arg == "all" then
 		Clear("props")
 		return
 	end
 
 	local steamid64 = false
 
-	if tonumber(arg) and ply:IsAdmin() then
+	if ply:IsAdmin() and tonumber(arg) then
 		steamid64 = tostring(arg)
 	end
 
 	if arg == "self" then
 		steamid64 = ply:SteamID64()
+		if CLIENT then
+			ClearPlayer("props")
+			print("Props list reset! Might not show until map change.")
+		end
 	end
 
 	if not steamid64 then
@@ -348,11 +356,9 @@ concommand.Add( "jazz_reset_props", function( ply, _, args )
 		return
 	end
 
-	local result = ClearPlayer("props", steamid64)
-	if CLIENT and result then
-		print("Props list successfully reset for " .. arg)
-	else
-		print("Something went wrong!")
+	if SERVER then
+		ClearPlayer("props", steamid64)
+		return
 	end
 
 end, nil, ResetHelpMsg)
